Shuffle definitions around and turn warnings into errors
